document_hierarchy_understanding_agentic_ai/requirements.txt
pyautogen>=0.2.0
anthropic>=0.40.0
chromadb>=0.4.0
